[추가] 학과 검색 알고리즘 구현 및 분리
